+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 ### Changed
 - Update dependencies.
 - Change folder structure.
+- Develop new database model.
+- Update add command.
+### Removed
+- Category.
+- Field.